ERIK/file_uploading/requirement.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
mysqlclient==2.1.1
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
